Moved Hooks Into One File
