--- codewit/client/src/hooks/modulehooks/useDeleteResources.tsx
@@ -1,17 +1,0 @@
-import axios from 'axios';
-
-const useDeleteResource = () => {
-
-  const deleteResource = async (resourceID: number | undefined) => {
-    try {
-      const response = await axios.delete(`/resources/${resourceID}`);
-      return response.data;
-    } catch (error) {
-      throw Error();
-    } 
-  };
-
-  return { deleteResource };
-};
-
-export default useDeleteResource;
