--- codewit/client/src/hooks/modulehooks/useFetchResources.tsx
@@ -1,18 +1,0 @@
-import { Resource } from '@codewit/interfaces'; 
-import axios from 'axios';
-
-const useFetchResources = () => {
-
-  const fetchResources = async () => {
-    try {
-      const response = await axios.get('/resources');
-      return response.data as Resource[];
-    } catch (error) {
-      throw Error();
-    } 
-  };
-
-  return { fetchResources };
-};
-
-export default useFetchResources;
